# internal/platform/apidocs/swagger.yaml
definitions:
...
  apierr.Envelope:
    properties:
      error:
        $ref: '#/definitions/apierr.EnvelopeEr'
      request_id:
        example: req_example
        type: string
      type:
        enum:
        - error
        example: error
        type: string
    required:
    - error
    - request_id
    - type
    type: object
  apierr.EnvelopeEr:
    properties:
      details:
        additionalProperties: {}
        type: object
      message:
        example: input is required
        type: string
      type:
        enum:
        - invalid_request_error
        - authentication_error
        - permission_error
        - not_found_error
        - request_too_large
        - rate_limit_error
        - api_error
        - timeout_error
        - overloaded_error
        example: invalid_request_error
        type: string
    required:
    - message
    - type
    type: object
...
  environments.CreateRequest:
    properties:
      name:
        type: string
      packages:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
        x-nullable: true
    required:
    - name
    type: object
  environments.Environment:
    properties:
      archived:
        type: boolean
      created_at:
        format: date-time
        type: string
      id:
        type: string
      name:
        type: string
      packages:
        additionalProperties:
          items:
            type: string
          type: array
        type: object
        x-nullable: true
    required:
    - archived
    - created_at
    - id
    - name
    - packages
    type: object
  environments.ListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/environments.Environment'
        type: array
      next_offset:
        description: 仅有下一页时返回；省略表示末页。 || Returned only when another page exists; omitted
          on the last page.
        type: integer
    required:
    - data
    type: object
...
info:
  contact: {}
  description: |-
    Managed Agents 服务。所有 /v1 接口使用 Bearer Wave API Key 认证； || Managed Agents service. All /v1 endpoints require a Bearer Wave API Key;
    由 `wave bootstrap` 签发。Authorization 填写完整的 `Bearer <Wave API Key>`。 || issue it with `wave bootstrap`. Set Authorization to the full `Bearer <Wave API Key>` value.
    offset 分页返回 data 和可选 next_offset（省略表示末页）；消息历史使用 next_after，事件列表使用最后一条 sequence 续读。 || Offset pagination returns data and an optional next_offset (omitted on the last page); message history uses next_after, and event lists resume from the last sequence.
    JSON 错误返回 {"type":"error","error":{"type":"api_error","message":"..."},"request_id":"req_..."}。文件传输的 Range 错误返回 text/plain；条件下载和健康检查可返回空响应。 || JSON errors return {"type":"error","error":{"type":"api_error","message":"..."},"request_id":"req_..."}. File-transfer Range errors return text/plain; conditional downloads and health checks may return empty responses.
    JSON 请求拒绝未知字段，非 multipart 请求体上限为 2 MiB。可空字段通过 x-nullable 标记；省略与 null 不等价。 || JSON requests reject unknown fields; non-multipart request bodies are limited to 2 MiB. Nullable fields use x-nullable; omission is not equivalent to null.
    交互式文档见 /swagger/index.html，OpenAPI spec 见 /swagger/doc.json。 || Interactive documentation: /swagger/index.html. OpenAPI spec: /swagger/doc.json.
  title: Wave AI API
  version: "1.0"
paths:
...
  /v1/environments:
    get:
      description: 分页返回当前所有者可见的环境配置。 || Return environment configurations visible
        to the current owner, with pagination.
      operationId: environmentsList
      parameters:
      - default: 100
        description: 每页条数 || Items per page
        in: query
        maximum: 200
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: 偏移量 || Offset
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/environments.ListResponse'
        "400":
          description: 分页参数不合法 || Invalid pagination parameters
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "401":
          description: 缺少或无效的 Bearer Key || Missing or invalid Bearer Key
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "403":
          description: Key 不具备 API scope || Key does not have API scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "500":
          description: 内部错误，可用 request_id 排查 || Internal error; use request_id for
            troubleshooting
          schema:
            $ref: '#/definitions/apierr.Envelope'
      security:
      - BearerAuth: []
      summary: 列出环境 || List environments
      tags:
      - environments
    post:
      consumes:
      - application/json
      description: 为当前所有者创建一个环境配置，声明各包管理器（apt、npm、pip、go、cargo、gem）要安装的包。 || Create
        an environment configuration for the current owner, declaring packages to
        install with apt, npm, pip, go, cargo and gem.
      operationId: environmentsCreate
      parameters:
      - description: 环境名称与包清单 || Environment name and package list
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/environments.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/environments.Environment'
        "400":
          description: 请求体不是合法 JSON、包管理器不受支持或包名为空 || Invalid JSON body, unsupported
            package manager or empty package name
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "401":
          description: 缺少或无效的 Bearer Key || Missing or invalid Bearer Key
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "403":
          description: Key 不具备 API scope || Key does not have API scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "500":
          description: 内部错误，可用 request_id 排查 || Internal error; use request_id for
            troubleshooting
          schema:
            $ref: '#/definitions/apierr.Envelope'
      security:
      - BearerAuth: []
      summary: 创建环境 || Create an environment
      tags:
      - environments
  /v1/environments/{id}:
    get:
      description: 按 ID 返回单个环境配置。 || Return an environment configuration by ID.
      operationId: environmentsGet
      parameters:
      - description: 环境 ID || Environment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/environments.Environment'
        "401":
          description: 缺少或无效的 Bearer Key || Missing or invalid Bearer Key
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "403":
          description: Key 不具备 API scope || Key does not have API scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "404":
          description: 环境不存在或不在当前所有者范围内 || Environment not found or outside the current
            owner's scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "500":
          description: 内部错误，可用 request_id 排查 || Internal error; use request_id for
            troubleshooting
          schema:
            $ref: '#/definitions/apierr.Envelope'
      security:
      - BearerAuth: []
      summary: 获取环境 || Get an environment
      tags:
      - environments
  /v1/environments/{id}/archive:
    post:
      description: 将环境标记为 archived，归档后不能再被部署引用。 || Mark the environment as archived;
        deployments can no longer reference it.
      operationId: environmentsArchive
      parameters:
      - description: 环境 ID || Environment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: 已归档 || Archived
        "401":
          description: 缺少或无效的 Bearer Key || Missing or invalid Bearer Key
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "403":
          description: Key 不具备 API scope || Key does not have API scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "404":
          description: 环境不存在或不在当前所有者范围内 || Environment not found or outside the current
            owner's scope
          schema:
            $ref: '#/definitions/apierr.Envelope'
        "500":
          description: 内部错误，可用 request_id 排查 || Internal error; use request_id for
            troubleshooting
          schema:
            $ref: '#/definitions/apierr.Envelope'
      security:
      - BearerAuth: []
      summary: 归档环境 || Archive an environment
      tags:
      - environments
...
securityDefinitions:
  BearerAuth:
    description: 填写完整的 Bearer <Wave API Key>，包含 Bearer 前缀和空格。 || Enter the full Bearer
      <Wave API Key> value, including the Bearer prefix and a space.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"